requirements.txt
streamlit>=1.31
langchain
langchain-openai
python-dotenv
pypdf
faiss-cpu
pandas
nest-asyncio

sentence-transformers>=2.2.0
langchain-community>=0.2.0
torch>=2.0.0
tenacity>=8.0.0

boto3>=1.34.0

